+++ database/migrations/2024_12_24_110134_create_temp_detail_retur_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('id_barang');
             $table->string('no_nota');
             $table->integer('qty');
-            $table->decimal('harga', 15, 2);
+            $table->double('harga');
             $table->timestamps();
         });
     }
